Add files via upload
This is the first addition to this repository, and I admit to you (reader) that there is a whole lot I do not understand -- about the GitHub processes and clean organization of public repositories. Time & iteration will clean this up, for now, alas, a trial by fire is the way.